YESS fixed paging again!

+++ src/kernel/memory/page.h
@@ -2,15 +2,15 @@
 
 typedef struct page
 {
-    unsigned int present    : 1;
-    unsigned int rw         : 1;
-    unsigned int user       : 1;
-    unsigned int reserved   : 2;
-    unsigned int accessed   : 1;
-    unsigned int dirty      : 1;
-    unsigned int reserved2  : 2;
-    unsigned int available  : 3;
-    unsigned int frame      : 20;
+    size_t present    : 1;
+    size_t rw         : 1;
+    size_t user       : 1;
+    size_t reserved   : 2;
+    size_t accessed   : 1;
+    size_t dirty      : 1;
+    size_t reserved2  : 2;
+    size_t available  : 3;
+    size_t frame      : 20;
 } page_t;
 
 typedef struct page_table
@@ -33,6 +33,7 @@ typedef struct page_directory_entry
     size_t available  : 3;
     size_t frame      : 20;
 } page_directory_entry_t;
+
 
 typedef struct page_directory
 {
